added patient details and get it

+++ CommonLayer/Models/UserModel.cs
@@ -6,7 +6,7 @@ namespace CommonLayer.Models
 {
     public class UserModel
     {
-        public long UserID { get; set; }
+        public int UserID { get; set; }
         public string FullName { get; set; }
         public string EmailID { get; set; }
         public string Password { get; set; }
